+++ Modules/ObjectiveTemplates.lua
@@ -32,11 +32,11 @@ function addon:CreateObjectiveTemplate(objectiveTitle, overwrite, supressSelect)
     ------------------------------------------------------------
     --Debug-----------------------------------------------------
     ------------------------------------------------------------
-    if objectiveTemplate.title == newObjectiveTitle then
-        print(format("DEBUG: Template successfully created: %s", newObjectiveTitle))
-    else
-        print(format("DEBUG: There was an error creating: %s", newObjectiveTitle))
-    end
+    -- if objectiveTemplate.title == newObjectiveTitle then
+    --     print(format("DEBUG: Template successfully created: %s", newObjectiveTitle))
+    -- else
+    --     print(format("DEBUG: There was an error creating: %s", newObjectiveTitle))
+    -- end
     ------------------------------------------------------------
     ------------------------------------------------------------