+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 3.0-alpha12
 - Implemented tracker progress alerts for multi-tracker objectives
 - Alert formats have been changed and now accept a user defined function returning a string
+- Bug fixes:
+  - -- Fixed error when deleting objective template
 
 ### 3.0-alpha11
 - Cleanup code