+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Farming Bar
 ## Version 3.0
+### 3.0-alpha13
+- Bug fixes:
+  - Removed "BAG_UPDATE_COOLDOWN" event from buttons to help performance issues
+  - Fixed error when dropping an objective from the DragFrame with not mouse focus
+  - Added appropriate data for tracker alerts to work with custom conditions
+
 ### 3.0-alpha12
 - You can now enable or disable alerts (customization not yet implemented)
 - Alert formats have been changed and now accept a user defined function returning a string (customization not yet implemented)